+++ Tests/MapboxMapsTests/Location/LocationManagerTests.swift
@@ -1,37 +1,31 @@
 import XCTest
 @testable import MapboxMaps
 
-internal class LocationManagerTests: XCTestCase {
+final class LocationManagerTests: XCTestCase {
 
-    var locationSupportableMapView: LocationSupportableMapViewMock!
-    var locationSupportableStyle: MockLocationStyleDelegate!
+    var style: MockLocationStyle!
 
     override func setUp() {
-        locationSupportableMapView = LocationSupportableMapViewMock()
-        locationSupportableStyle = MockLocationStyleDelegate()
         super.setUp()
+        style = MockLocationStyle()
     }
 
     override func tearDown() {
-        locationSupportableMapView = nil
-        locationSupportableStyle = nil
+        style = nil
         super.tearDown()
     }
 
     func testLocationManagerDefaultInitialization() {
         let locationOptions = LocationOptions()
 
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
 
         XCTAssertEqual(locationManager.options, locationOptions)
-        XCTAssertTrue(locationManager.locationSupportableMapView === locationSupportableMapView)
         XCTAssertNil(locationManager.delegate)
     }
 
     func testAddLocationConsumer() {
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
         let locationConsumer = LocationConsumerMock()
 
         locationManager.addLocationConsumer(newConsumer: locationConsumer)
@@ -42,8 +36,7 @@ internal class LocationManagerTests: XCTestCase {
     func testUpdateLocationOptionsWithModifiedPuckType() {
         var locationOptions = LocationOptions()
         locationOptions.puckType = .puck2D(Puck2DConfiguration(scale: .constant(1.0)))
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
 
         var locationOptions2 = LocationOptions()
         locationOptions2.puckType = .puck2D(Puck2DConfiguration(scale: .constant(2.0)))
@@ -56,8 +49,7 @@ internal class LocationManagerTests: XCTestCase {
     func testUpdateLocationOptionsWithPuckTypeSetToNil() {
         var locationOptions = LocationOptions()
         locationOptions.puckType = .puck2D()
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
 
         var locationOptions2 = LocationOptions()
         locationOptions2.puckType = nil
@@ -70,8 +62,7 @@ internal class LocationManagerTests: XCTestCase {
     func testUpdateLocationOptionsWithPuckTypeSetToNonNil() {
         var locationOptions = LocationOptions()
         locationOptions.puckType = nil
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
 
         var locationOptions2 = LocationOptions()
         locationOptions2.puckType = .puck2D()
@@ -84,8 +75,7 @@ internal class LocationManagerTests: XCTestCase {
     func testUpdateLocationOptionsWithCoursePuckBearingSource() {
         var locationOptions = LocationOptions()
         locationOptions.puckType = .puck2D()
-        let locationManager = LocationManager(locationSupportableMapView: locationSupportableMapView,
-                                              style: locationSupportableStyle)
+        let locationManager = LocationManager(style: style)
 
         locationManager.options = locationOptions
         XCTAssertEqual(locationManager.locationPuckManager?.puckBearingSource, .heading)