+++ Tests/MapboxMapsTests/Location/MockLocationStyle.swift
@@ -1,7 +1,7 @@
 import Foundation
 @testable import MapboxMaps
 
-class MockLocationStyleDelegate: LocationStyleDelegate {
+final class MockLocationStyle: LocationStyleProtocol {
     func _addPersistentLayer(_ layer: Layer, layerPosition: LayerPosition?) throws {}
     func removeLayer(withId id: String) throws {}
     func layerExists(withId id: String) -> Bool { return false }